% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_hv.R
\name{image_hv}
\alias{image_hv}
\title{Download images form Herbário Virtual Reflora}
\usage{
image_hv(scientific.name, tx = 1, destfolder = "images",
  municipality = NULL)
}
\arguments{
\item{scientific.name}{A character string with the scientific name.}

\item{tx}{Multiplication rate of standard image size  (70 x 150 px).}

\item{destfolder}{A character string with the folder where the images will be saved}

\item{municipality}{A character string with the municipality name. If is NULL (default) download images of all available municipalities.}
}
\value{
A rasterStack with the minimun, maximum, median, mean and standard
}
\description{
Download images form Herbário Virtual Reflora
}
\examples{
image_hv("Prepusa montana")

}
\seealso{
(<http://reflora.jbrj.gov.br/reflora/herbarioVirtual/ConsultaPublicoHVUC/ConsultaPublicoHVUC.do>)
}
