% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_hv.R
\name{image_hv}
\alias{image_hv}
\title{Download images form Herbário Virtual Reflora}
\usage{
image_hv(fam, gen, epi, tx = 1)
}
\arguments{
\item{fam}{A character string with the family name}

\item{gen}{A character string with the genus name}

\item{epi}{A character string with the specific epithet name}
}
\value{
A rasterStack with the minimun, maximum, median, mean and standard
}
\description{
Download images form Herbário Virtual Reflora
}
\examples{
fam = "Melastomataceae"
gen = "Tibouchina"
epi = "clavata"
image_hv(fam,gen,epi)

}
\seealso{
\link{'http://reflora.jbrj.gov.br/reflora/herbarioVirtual'}
}
